+++ apps/webview/app/mashong/[team]/_components/PopcornXpTracker.tsx
@@ -7,8 +7,8 @@ import { useState } from 'react';
 import { toast } from 'react-hot-toast/headless';
 import { assert, isKeyOfObject } from 'utils';
 
-import { feedPopcorn } from '@/app/_actions/feedPopcorn';
 import { showErrorToast } from '@/app/_components/ErrorToast';
+import { feedPopcorn } from '@/app/mashong/_actions/feedPopcorn';
 import { css } from '@/styled-system/css';
 import SvgImage from '@/ui/svg-image';
